* commit '5ced907f129ac7739841948bea56c54b5e4ebf16': (380 commits)
  无jira任务, sonar问题
  CHART-11446 未选择数据集灰化了combox
  CHART-11446 未选择数据集灰化了combox
  CHART-11242 标题界面
  CHART-11513 数据集配置面板布局优化
  sonar问题
  CHART-11424 fix 补充几种图表单元格数据面板 数据筛选组件
  fix 雷达图轴面板NPE
  CHART-11424 fix 数据集面板-数据筛选
  CHART-11093 fix 数据面板切换
  代码质量
  有些层级可能没配数据
  REPORT-23978 代码合并出错
  REPORT-20330 sonar问题修复 常量放最上面
  REPORT-20330 sonar问题修复 修复 sonar 问题
  CHART-10227 地图 钻取地图 热力地图 数据配置界面
  CHART-11093 不实例化，类型不同就不populate 验证了没问题
  CHART-11093 反射获取相应数据面板对应的数据类型
  CHART-11321 fix 坐标轴显示策略面板checkUse
  CHART-11093 refactor 数据面板切换
  ...

+++ designer-base/src/main/java/com/fr/design/layout/TableLayoutHelper.java
@@ -372,7 +372,8 @@ public class TableLayoutHelper {
         double p = TableLayout.PREFERRED;
         double f = TableLayout.FILL;
         JPanel jp1 = TableLayoutHelper.createTableLayoutPane(createTestComponents("jp1"), TableLayoutHelper.FILL_NONE);
-        JPanel jp2 = TableLayoutHelper.createGapTableLayoutPane(createTestComponents("jp2"), TableLayoutHelper.FILL_LASTCOL_AND_ROW, 2 * TEN, 2 * TEN);
+        JPanel jp2 = TableLayoutHelper.createGapTableLayoutPane(createTestComponents("jp2"),
+                TableLayoutHelper.FILL_LASTCOL_AND_ROW, (double)2 * TEN, (double)2 * TEN);
         JPanel jp3 = TableLayoutHelper.createGapTableLayoutPane(createTestComponents("jp3"),
                 new double[]{f, p, f, p}, new double[]{f, f}, 4, 4);
         JPanel jp4 = TableLayoutHelper.createGapTableLayoutPane(createTestComponents("jp4"),
